refactor(dashboard): extract subscription lifecycle tracking from webhook handler
`subscription-event.ts` is back to handler-sized — it does its primary job
(toggling Unkey access) and delegates analytics to the new
`subscription-lifecycle-tracking.ts`. Behavior is unchanged; this only
separates concerns so the file matches the size/shape of the other
webhook handlers (customer-event, invoice-event).
